Basic Pages created
I have also created a few variables within the JavaScript to start everything off. I have commented what these variables should be used for or what they refer to.

the main CSS page has some standard settings for windows within all pages using innerPageContent1, 2 and 3. Which create the outer div container, the darker background and the smaller inner div of a lighter color.

The navigation menu is currently in the header and centered. I figured we could possibly have it inside of a hamburger menu.

Colors are done with variables and can easily be adjusted, accent colors added, borders etc.